requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.2.0
langchain-community==0.2.0
langchain-google-genai==1.0.10
faiss-cpu==1.8.0.post1
sentence-transformers==2.7.0
python-dotenv==1.0.0
pandas==2.1.3
deep-translator==1.11.4
langdetect==1.0.9
pydantic==2.5.2
google-generativeai==0.7.2
